fix: base dialog now emits escape event with onKeyStroke, fix avatars, remove crud forms
