+++ src/components/Generic/Dialogs/PrinterDialog.vue
@@ -1,15 +1,15 @@
 <template>
   <BaseDialog
     :id="dialog.dialogId"
-    :dialogCloseEvent="clearForm"
     :max-width="showChecksPanel ? '900px' : '600px'"
+    @escape="closeDialog()"
   >
     <validation-observer ref="validationObserver" v-slot="{ invalid }">
       <v-card>
         <v-card-title>
           <span class="text-h5">
             <v-avatar color="primary" size="56">
-              {{ avatarInitials() }}
+              {{ avatarInitials }}
             </v-avatar>
             <span v-if="isUpdating"> Updating Printer </span>
             <span v-else> New Printer </span>
@@ -18,7 +18,93 @@
         <v-card-text>
           <v-row>
             <v-col :cols="showChecksPanel ? 8 : 12">
-              <PrinterCrudForm ref="printerCrudForm" :printer-id="storedPrinter?.id" />
+              <v-container>
+                <v-row>
+                  <v-col v-if="formData" cols="12" md="6">
+                    <validation-provider v-slot="{ errors }" :rules="printerNameRules" name="Name">
+                      <v-text-field
+                        v-model="formData.printerName"
+                        :counter="printerNameRules.max"
+                        :error-messages="errors"
+                        autofocus
+                        label="Printer name*"
+                        required
+                      />
+                    </validation-provider>
+
+                    <validation-provider
+                      v-slot="{ errors }"
+                      name="Printer IP or HostName"
+                      rules="required|ip_or_fqdn"
+                    >
+                      <v-text-field
+                        v-model="formData.printerHostName"
+                        :error-messages="errors"
+                        hint="Examples: 'my.printer.com', 'localhost' or '192.x.x.x'"
+                        label="IP/Host*"
+                      ></v-text-field>
+                    </validation-provider>
+                  </v-col>
+                  <v-col cols="12" md="6">
+                    <validation-provider v-slot="{ errors }" name="Enabled">
+                      <v-checkbox
+                        v-model="formData.enabled"
+                        :error-messages="errors"
+                        hint="Disabling makes the printer passive"
+                        label="Enabled*"
+                        persistent-hint
+                        required
+                      ></v-checkbox>
+                    </validation-provider>
+
+                    <validation-provider
+                      v-slot="{ errors }"
+                      name="Host Port"
+                      rules="required|integer|max:65535"
+                    >
+                      <v-text-field
+                        v-model="formData.printerHostPort"
+                        :error-messages="errors"
+                        hint="Examples: '80', '443' or '5050'"
+                        label="Host Port*"
+                      ></v-text-field>
+                    </validation-provider>
+                  </v-col>
+                  <v-col class="pb-5 pt-0" cols="12" md="12">
+                    <validation-provider v-slot="{ errors }" :rules="apiKeyRules" name="ApiKey">
+                      <v-text-field
+                        v-model="formData.apiKey"
+                        :counter="apiKeyRules.length"
+                        :error-messages="errors"
+                        hint="User or Application Key only (Global API key fails)"
+                        label="API Key*"
+                        persistent-hint
+                        required
+                      ></v-text-field>
+                    </validation-provider>
+                  </v-col>
+                </v-row>
+
+                <v-expansion-panels accordion>
+                  <v-expansion-panel>
+                    <v-expansion-panel-header>Advanced settings</v-expansion-panel-header>
+                    <v-expansion-panel-content>
+                      <v-col cols="12" md="12">
+                        <validation-provider v-slot="{ errors }" name="PrinterHostPrefix">
+                          <v-select
+                            v-model="formData.printerHostPrefix"
+                            :error-messages="errors"
+                            :items="['http', 'https']"
+                            label="Insecure/Secure HTTP"
+                            required
+                            value="http"
+                          ></v-select>
+                        </validation-provider>
+                      </v-col>
+                    </v-expansion-panel-content>
+                  </v-expansion-panel>
+                </v-expansion-panels>
+              </v-container>
             </v-col>
 
             <PrinterChecksPanel v-if="showChecksPanel" :cols="4">
@@ -52,8 +138,8 @@
             Test connection
           </v-btn>
 
-          <v-btn :disabled="invalid" color="blue darken-1" text @click="submit()"
-            >{{ submitButtonText }}
+          <v-btn :disabled="invalid" color="blue darken-1" text @click="submit()">
+            {{ submitButtonText }}
           </v-btn>
         </v-card-actions>
       </v-card>
@@ -62,29 +148,36 @@
 </template>
 
 <script lang="ts">
-import { defineComponent } from "vue";
-import { ValidationObserver } from "vee-validate";
+import { defineComponent, inject } from "vue";
+import { ValidationObserver, ValidationProvider } from "vee-validate";
 import { generateInitials } from "../../../shared/noun-adjectives.data";
 import { infoMessageEvent } from "../../../shared/alert.events";
 import { usePrinterStore } from "../../../store/printer.store";
 import { PrintersService } from "@/backend";
 import PrinterChecksPanel from "@/components/Generic/Dialogs/PrinterChecksPanel.vue";
-import PrinterCrudForm from "@/components/Generic/Forms/PrinterCrudForm.vue";
 import { DialogName } from "@/components/Generic/Dialogs/dialog.constants";
 import { useTestPrinterStore } from "../../../store/test-printer.store";
-import { CreatePrinter } from "@/models/printers/crud/create-printer.model";
+import {
+  CreatePrinter,
+  getDefaultCreatePrinter,
+  PreCreatePrinter,
+} from "@/models/printers/crud/create-printer.model";
 import { useDialog } from "../../../shared/dialog.composable";
+import { AppConstants } from "../../../shared/app.constants";
+
+const watchedId = "printerId";
 
 interface Data {
   showChecksPanel: boolean;
   copyPasteConnectionString: string;
+  formData: PreCreatePrinter;
 }
 
 export default defineComponent({
   name: "PrinterDialog",
   components: {
+    ValidationProvider,
     ValidationObserver,
-    PrinterCrudForm,
     PrinterChecksPanel,
   },
   setup: () => {
@@ -93,16 +186,26 @@ export default defineComponent({
       printersStore: usePrinterStore(),
       testPrinterStore: useTestPrinterStore(),
       dialog,
+      appConstants: inject("appConstants") as AppConstants,
     };
   },
-  async created() {},
+  async created() {
+    if (this.printerId) {
+      const crudeData = this.printersStore.printer(this.printerId) as CreatePrinter;
+      this.formData = PrintersService.convertPrinterToCreateForm(crudeData);
+    }
+  },
   async mounted() {},
   props: {},
   data: (): Data => ({
     showChecksPanel: false,
     copyPasteConnectionString: "",
+    formData: getDefaultCreatePrinter(),
   }),
   computed: {
+    printerId() {
+      return this.printersStore.updateDialogPrinter?.id;
+    },
     storedPrinter() {
       return this.printersStore.updateDialogPrinter;
     },
@@ -124,11 +227,27 @@ export default defineComponent({
     isClipboardApiAvailable() {
       return navigator.clipboard;
     },
-    printerCrudForm() {
-      return this.$refs.printerCrudForm as InstanceType<typeof PrinterCrudForm>;
+    avatarInitials() {
+      if (this.formData) {
+        return generateInitials(this.formData?.printerName);
+      }
+      return "?";
+    },
+    printerNameRules() {
+      return { required: true, max: this.appConstants.maxPrinterNameLength };
+    },
+    apiKeyRules() {
+      return {
+        required: true,
+        length: this.appConstants.apiKeyLength,
+        alpha_num: true,
+      };
     },
   },
   methods: {
+    resetForm() {
+      this.formData = getDefaultCreatePrinter();
+    },
     async quickCopyConnectionString() {
       const printer = this.storedPrinter;
       if (!printer) return;
@@ -148,24 +267,15 @@ export default defineComponent({
       }
       await navigator.clipboard.writeText(connectionString);
     },
-    formData() {
-      return this.printerCrudForm?.formData;
-    },
-    avatarInitials() {
-      const formData = this.formData();
-      if (formData) {
-        return generateInitials(formData?.printerName);
-      }
-    },
+
     openTestPanel() {
       this.showChecksPanel = true;
     },
     async testPrinter() {
       if (!(await this.isValid())) return;
-      const formData = this.formData();
-      if (!formData) return;
+      if (!this.formData) return;
 
-      const testPrinter = PrintersService.convertCreateFormToPrinter(formData);
+      const testPrinter = PrintersService.convertCreateFormToPrinter(this.formData);
       if (!testPrinter) return;
       this.openTestPanel();
 
@@ -174,8 +284,7 @@ export default defineComponent({
       this.testPrinterStore.currentCorrelationToken = correlationToken;
     },
     async pasteFromClipboardOrField() {
-      const formData = this.formData();
-      if (!formData) return;
+      if (!this.formData) return;
 
       if (!this.isClipboardApiAvailable && !this.copyPasteConnectionString?.length) {
         return;
@@ -186,7 +295,7 @@ export default defineComponent({
         : this.copyPasteConnectionString;
       const printerObject = JSON.parse(jsonData);
 
-      PrintersService.applyLoginDetailsPatchForm(printerObject, formData);
+      PrintersService.applyLoginDetailsPatchForm(printerObject, this.formData);
     },
     async isValid() {
       return await this.validationObserver.validate();
@@ -207,9 +316,8 @@ export default defineComponent({
     },
     async submit() {
       if (!(await this.isValid())) return;
-      const formData = this.formData();
-      if (!formData) return;
-      const createPrinter = PrintersService.convertCreateFormToPrinter(formData);
+      if (!this.formData) return;
+      const createPrinter = PrintersService.convertCreateFormToPrinter(this.formData);
       if (this.isUpdating) {
         await this.updatePrinter(createPrinter);
       } else {
@@ -218,19 +326,19 @@ export default defineComponent({
       this.closeDialog();
     },
     closeDialog() {
+      this.dialog.closeDialog();
       this.showChecksPanel = false;
       this.testPrinterStore.clearEvents();
-      this.printerCrudForm?.resetForm();
-      this.dialog.closeDialog();
+      this.resetForm();
       this.printersStore.updateDialogPrinter = undefined;
       this.copyPasteConnectionString = "";
     },
-    clearForm() {
-      this.showChecksPanel = false;
-      this.testPrinterStore.clearEvents();
-      this.printerCrudForm?.resetForm();
-      this.printersStore.updateDialogPrinter = undefined;
-      this.copyPasteConnectionString = "";
+  },
+  watch: {
+    [watchedId](val?: string) {
+      if (!val) return;
+      const printer = this.printersStore.printer(val) as CreatePrinter;
+      this.formData = PrintersService.convertPrinterToCreateForm(printer);
     },
   },
 });
